fix: Implement proper dark/light mode theme switching across all pages
- Added theme initialization script to prevent FOUC (Flash of Unstyled Content)

- Updated all main pages (index, reservations, billing, guests, analytics, housekeeping, rooms-overview, login, setup-2fa, verify-2fa)

- Theme preference is now stored in localStorage and persists across sessions

- Theme applies immediately on page load before any content renders

- Updated page titles from 'Core 1 PMS' to 'Inn Nexus Hotel Management System'

+++ includes/db.php
@@ -113,9 +113,9 @@ function verifyTOTPCode($userIdOrSecret, string $code): bool {
     $secret = $userIdOrSecret;
   }
   
-  $timeSlice = floor(time() / 30);
+  $timeSlice = (int)floor(time() / 30);
   for ($i = -1; $i <= 1; $i++) {
-    $calculatedCode = calculateTOTP($secret, $timeSlice + $i);
+    $calculatedCode = calculateTOTP($secret, (int)($timeSlice + $i));
     if (hash_equals($calculatedCode, $code)) {
       return true;
     }
